Backend/Gateway/requirements.txt
asgiref==3.7.2
Django==4.1
django-environ==0.11.2
djangorestframework==3.14.0
pytz==2023.3.post1
setuptools==68.2.2
sqlparse==0.4.4
wheel==0.41.3
